Moved docker particulars to the root of project, which helps to unify dev/test/prod environments. Using docker-compose config to generate a full compose file with env vars. Updated pycord. Updated permissions checks to new API. Entrypoint now waits 20secs for DB to start.
