+++ docker-compose.yaml
@@ -1,10 +1,10 @@
 version: "3.3"
 services:
   discord-bot:
-    image: infochallengebot:latest
+    image: gregjan/infochallengebot:0.6
     build:
-      context: ../
-      dockerfile: docker/Dockerfile
+      context: .
+      dockerfile: Dockerfile
     environment:  # cofigure these in your .env file outside of container
       - IS_PROD=${IS_PROD}
       - BOT_TOKEN=${BOT_TOKEN}
